refactoring, makefile, version info
