+++ cmd/client/sync.go
@@ -1,4 +1,4 @@
-package cmd
+package main
 
 import "github.com/spf13/cobra"
